requirements.txt
requests
marathon
docker-2.0.0
envtpl
PyYAML
kazoo
kafka-python
click
